Change #define macro to static function and move functions to another file.
