+++ browser/ui/BUILD.gn
@@ -257,6 +257,8 @@ source_set("ui") {
       "views/tabs/brave_tab_prefs.h",
       "views/tabs/brave_tab_strip.cc",
       "views/tabs/brave_tab_strip.h",
+      "views/tabs/brave_tab_strip_layout_helper.cc",
+      "views/tabs/brave_tab_strip_layout_helper.h",
       "views/tabs/features.cc",
       "views/tabs/features.h",
       "views/toolbar/bookmark_button.cc",
